Refactored utils.
