+++ ipymd/formats/tests/_utils.py
@@ -52,32 +52,6 @@ def _read_test_file(basename, format):
     return _load_file(path, format)
 
 
-#------------------------------------------------------------------------------
-# Test Markdown parser
-#------------------------------------------------------------------------------
-
-
-def _diff_removed_lines(diff):
-    return ''.join(x[2:] for x in diff if x.startswith('- '))
-
-
-def _diff(text_0, text_1):
-    """Return a diff between two strings."""
-    diff = difflib.ndiff(text_0.splitlines(), text_1.splitlines())
-    return _diff_removed_lines(diff)
-
-
-def _show_outputs(*outputs):
-    for output in outputs:
-        print()
-        print("-" * 30)
-        pprint(output)
-
-
-#------------------------------------------------------------------------------
-# Test Markdown parser
-#------------------------------------------------------------------------------
-
 def _test_reader(basename, format):
     """Return converted and expected ipymd cells of a given example."""
     contents = _read_test_file(basename, format)
