+++ ipymd/formats/tests/test_atlas.py
@@ -7,8 +7,8 @@
 #------------------------------------------------------------------------------
 
 from ...core.format_manager import format_manager, convert
-from ...utils.utils import _remove_output
-from ._utils import (_test_reader, _test_writer, _diff, _show_outputs,
+from ...utils.utils import _remove_output, _diff, _show_outputs
+from ._utils import (_test_reader, _test_writer,
                      _exec_test_file, _read_test_file)
