+++ ipymd/formats/tests/test_notebook.py
@@ -8,7 +8,8 @@
 
 from ...core.format_manager import format_manager, convert
 from ..notebook import _compare_notebooks
-from ._utils import (_test_reader, _test_writer, _diff, _show_outputs,
+from ...utils.utils import _diff, _show_outputs
+from ._utils import (_test_reader, _test_writer,
                      _exec_test_file, _read_test_file)
