+++ ipymd/formats/tests/test_utils.py
@@ -7,7 +7,7 @@
 #------------------------------------------------------------------------------
 
 import os.path as op
-from ._utils import _test_file_path, _exec_test_file, _diff
+from ._utils import _test_file_path, _exec_test_file
 
 
 #------------------------------------------------------------------------------
@@ -22,11 +22,3 @@ def test_file_path():
 def test_exec_test_file():
     filename = 'ex1'
     assert isinstance(_exec_test_file(filename), list)
-
-
-def test_diff():
-    s = 'abcdef ghijkl'
-    assert _diff(s, s) == ''
-
-    assert _diff(s, ' ' + s) == s
-    assert _diff(s, s + ' ') == s
