+++ ipymd/utils/tests/test_tempdir.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Tests of HDF5 routines."""
+"""Test tempdir."""
 
 #------------------------------------------------------------------------------
 # Imports
